Added functions to Get and Set MDData in OpenMMCubes/utils.py. The Get function returna a MDData dictionary with kewords: “structure” : Parmed structure “positions” : If present system current atom positions “topology“ : Parmed topology “box” : If present box vectors otherwise None “parameters” : Parmed force field parameters “velocities” : If present system atom velocities “ref_positions” : System reference atom positions
To the OEMol() passed between cubes, it has been added an OpenMM Quantity object which carries
the reference positions. The reference positions are set in the OpenMMCubes/cubes.py, immediately
after the complex creation.

In LigPrepCubes/cubes.py the keep_confs keyword in the function call ff_utils.assignCharges() has
been set to None to avoid that the ligand will be placed out of the binding site in the complex
creation.

In LigPrepCubes/utils.py the keywords net_charge and charge_method have been set to None to avoid to
recompute the ligand charges after the quacpac ligand charge calculations